+++ src/main/groovy/all/shared/gradle/file/FileListerPlugin.groovy
@@ -9,7 +9,21 @@ import org.gradle.api.Project
 
 @CompileStatic
 class FileListerPlugin implements Plugin<Project> {
+  static final String EXTENSION_NAME = 'fileLister'
+
+  static final boolean complement(final Project project) {
+    if (project.extensions.findByName(EXTENSION_NAME) == null) {
+      project.extensions.create(EXTENSION_NAME, FileListerExtension, project)
+      project.logger.debug('Added file-lister extension')
+      true
+    }
+    else {
+      project.logger.error('Couldn\'t add file-lister extension')
+      false
+    }
+  }
+
   void apply(final Project project) {
-    project.extensions.create('fileLister', FileLister, project)
+    complement(project)
   }
 }