+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # File Lister Change Log
+
+## 1.0.1 - October 2018
+
+* Adds two new static fields to the `FileListerPlugin`.
+* Renames `FileLister` to `FileListerExtension`.
+* Improves `build.gradle`.
+* Updates README file.